refactor: add reducers
