+++ src/core/events/draw/rectangle.ts
