+++ src/core/events/draw/square.ts
