+++ src/core/events/mouse/left.ts
