+++ src/core/reducers/index.ts
@@ -1,0 +1,3 @@
+export { drawReducer } from './draw-reducer'
+export { mouseReducer } from './mouse-reducer'
+export { screenReducer } from './screen-reducer'
