--- src/events/index.ts
@@ -1,1 +1,0 @@
-export { dispatchEvent } from './dispatch-event'
